Calculate gene set scores

Calculate gene set scores using the approach described in Tirosh2016

Usage

calculate_gs_scores(
  scMuffinList = NULL,
  gs_list = NULL,
  mc.cores = 2,
  nbins = 25,
  nmark_min = 5,
  ncells_min = 10,
  k = 100,
  kmin = 50,
  score_type = c("relative", "mean"),
  null_model = TRUE,
  verbose = TRUE,
  na.rm = TRUE,
  overwrite = FALSE
)

Arguments

scMuffinList

scMuffinList object

gs_list

list of gene sets

mc.cores

number of cores

nbins

number of bins to split the distribution of average gene expression

nmark_min

number of minimum markers that are required for the succesful calculation of a gene set score

ncells_min

number of minimum cells in which a gene set has to be succesfully calculated

k

number of permutations

kmin

minimum number of permutations; due to missing values it is hard to ensure that a gene set score can be compared to k permutations in every cell

score_type

type of score. if "relative", than the score is the difference between the observed gene set average expression and that of a k permutations; if "mean" the score is equal to the observed gene set average expression

null_model

TRUE if permutations have to be used. Required for score_type="relative"

verbose

verbosity

na.rm

whether to use NA or not

overwrite

whether to update or not gene_set_scoring and gene_set_scoring_full elements of scMuffinList.

Value

scMuffinList with element gene_set_scoring, a list that contains summary and full. The element summary contains a cells-by-gene sets data.frame. The element "full" contains a data.frame for each gene set. See [gs_score()] for further details.

References

Tirosh2016 10.1126/science.aad0501