Gene Set Enrichment Analysis

Gene Set Enrichment Analysis

Usage

csea(
  rl = NULL,
  gsl = NULL,
  k = 100,
  min.size = 100,
  ord.mode = -1,
  min.k = 10,
  mc_cores_path = 1,
  mc_cores_perm = 1
)

Arguments

rl

numeric matrix of genes-by-ranking criteria; each column contains numeric values; rownames are mandatory

gsl

named list of gene sets

k

number of permutations

min.size

minimum number of cells with a not null value

ord.mode

ordering mode: -1 -> descending; 1 ascending; must be of length equal to `ncol(rl)`

min.k

minimum number of valid permutations to support empirical nulls

mc_cores_path

number of cores to use for parallel calculation of gene set lists; the total number of cpu used will be mc_cores_path * mc_cores_perm

mc_cores_perm

number of cores to use for parallel calculation of ranked list permutations; the total number of cpu used will be mc_cores_path * mc_cores_perm

Value

list with two data.frames, gs_table and leading_edge.