Define a proliferation score

Define a proliferation score on the basis of G1/S and G2/M markers (by default gsls_Symbol$Tirosh)

Usage

proliferation_analysis(
  scMuffinList = NULL,
  mc.cores = 1,
  nbins = 25,
  nmark_min = 5,
  ncells_min = 5,
  k = 99,
  kmin = 49,
  score_type = c("relative", "mean"),
  null_model = TRUE,
  mean_scale = TRUE,
  gsl = NULL
)

Arguments

scMuffinList

scMuffinList object

mc.cores

number of cores

nbins

number of bins to split the distribution of average gene expression

nmark_min

number of minimum markers that are required for the succesful calculation of a signature

ncells_min

number of minimum cells in which a gene set has to be succesfully calculated

k

number of permutations

kmin

minimum number of permutations; due to missing values it is hard to ensure that a signature can be compared to k permutations in every cell

score_type

type of score. if "relative", then the score is the difference between the observed gene set average expression and that of a k permutations; if "mean" the score is equal to the observed gene set average expression

null_model

TRUE if permutations have to be used. Required for score_type="relative"

mean_scale

whether to scale the values obtained using score_type="mean"

gsl

list with two gene sets with 1/S and G2/M markers. If NULL gsls_Symbol$Tirosh is used. See gsls_Symbol$Tirosh to properly format custom gene sets.

Value

scMuffinList with element "proliferation", a list with:

summary: data.frame with proliferation value

full: data.frame with the vales of each of the two considered gene sets